Refactored initial code of resume
